pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymath-cli"
version = "0.1.0"
description = "CLI tools for systematic polymathic learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Alex"}
]
keywords = ["learning", "knowledge-management", "obsidian", "polymath"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
pm = "pm.cli:cli"
pm-init = "pm.commands.init:init"
pm-status = "pm.commands.status:status"
pm-next = "pm.commands.next_cmd:next_cmd"
pm-pair = "pm.commands.pair:pair"
pm-log = "pm.commands.log:log"
pm-gaps = "pm.commands.gaps:gaps"

[tool.setuptools.packages.find]
where = ["."]
include = ["pm*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
